+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for `mempack`
 
+## 0.1.1.0
+
+* Add `packWithByteArray` and `packWithMutableByteArray`
+* Fix infinite loop during list decoding when length encoded was negative
+
 ## 0.1.0.0
 
 * Initial release